refresh the sync script

+++ sync-to-llvm.sh
@@ -6,22 +6,38 @@ if test $WHOAMI != "jenkins"; then
 	echo "should be run under jenkins"
 fi
 
-if test $# -ne 1; then
+if test $# -ne 1 -a $# -ne 2; then
 	echo "Wrong number of args."
-	echo "Syntax: $0 <repository>"
+	echo "Syntax: $0 <repository> <skip sync>"
 	exit 1
 fi
 
+if ! grep -q XXX= ~/.ssh/known_hosts; then
+	echo "Key unknown. added"
+	echo "XXX" >> ~/.ssh/known_hosts
+fi
+
+
 REPOSITORY=$1
-TARGET=apt@apt-origin.llvm.org
+SKIP_SYNC=$2
+TARGET=XXXX@XXX.llvm.org
 BASE_TARGETDIR=/data/apt/www
 BASE_LOCALDIR=/srv/repository
 if test ! -d $BASE_LOCALDIR/$REPOSITORY; then
 	echo "Cannot find directory $REPOSITORY"
 	exit 1
 fi
-find $BASE_LOCALDIR -type d | xargs chmod 755
+
+LLVM_DEFAULT_DIR=$BASE_LOCALDIR/$REPOSITORY/pool/main/l/llvm-defaults/
+if test ! -d $LLVM_DEFAULT_DIR/; then
+        echo "Cannot find directory $LLVM_DEFAULT_DIR"
+        exit 1
+fi
+
+if test -z "$SKIP_SYNC"; then
+find $BASE_LOCALDIR -type d | xargs chmod 755 || true
 find $BASE_LOCALDIR -type f ! -name sync-to-llvm.sh | xargs -I {}  -d '\n' chmod 644 "{}" || true
+ssh $TARGET mkdir -p $BASE_TARGETDIR/$REPOSITORY
 echo "Delete potential old directory"
 time ssh $TARGET rm -rf $BASE_TARGETDIR/$REPOSITORY.back
 echo "Copy the current repo to a new directory to be updated"
@@ -34,8 +50,9 @@ echo "Move the new repo to the actual dir"
 time ssh $TARGET mv $BASE_TARGETDIR/$REPOSITORY.back $BASE_TARGETDIR/$REPOSITORY
 echo "Delete the old repo"
 time ssh $TARGET rm -rf $BASE_TARGETDIR/$REPOSITORY.1
+fi
 
-key="TO UDPATE"
+key="xxxx"
 if test "$REPOSITORY" == "unstable"; then
     REPOSITORY_CODE=""
 else
@@ -49,3 +66,5 @@ url="InRelease Release Release.gpg"
 for f in $url; do
 curl -XPOST -H "Fastly-Key:$key" https://api.fastly.com/purge/apt.llvm.org/$REPOSITORY/dists/llvm-toolchain$REPOSITORY_CODE/$f
 done
+curl -XPOST -H "Fastly-Key:$key" https://api.fastly.com/purge/apt.llvm.org/$REPOSITORY/
+
